+++ src/components/__tests__/Intl.test.jsx
@@ -1,0 +1,22 @@
+// Copyright (c) 2021 Antti Kivi
+// Licensed under the MIT License
+
+import React from 'react';
+import { render } from '@testing-library/react';
+import { useStaticQuery } from 'gatsby';
+
+import Intl from '../Intl';
+
+import intlQuery from '../../../test/data/intlQuery';
+
+describe('Intl component', () => {
+  beforeAll(() => useStaticQuery.mockReturnValue(intlQuery));
+
+  it('renders correctly', () => {
+    const { container, getByText } = render(<Intl locale="fi">Content</Intl>);
+
+    expect(container).toMatchSnapshot();
+
+    expect(getByText('Content')).toBeInTheDocument();
+  });
+});